pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    # "markdown>=3.9",  # Commented out - markdown support disabled
    "networkx>=3.5",
    "nltk>=3.9.2",
    "odfpy>=1.4.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.2",
    "spacy>=3.8.7",
    "streamlit>=1.50.0",
    "striprtf>=0.0.29",
]
